+++ src/phys/fracture/fracture_calculator/fracture_calculator.cpp
@@ -63,14 +63,14 @@ namespace pe_phys_fracture {
                                                                    const pe::Vector3 &p, const pe::Vector3 &n,
                                                                    FractureDataManager &new_mesh) {
         auto face = old_mesh.get_face(face_id);
-        uint32_t vert_count = face.vert_ids.size();
+        uint32_t vert_count = (uint32_t)face.vert_ids.size();
         utils::hash_vector<pe::Vector3> inter_points(FRAC_VEC_INIT(vert_count));
         pe::Array<vertex> vertices(vert_count);
         pe::Array<int> side(vert_count, -1);
         utils::hash_vector<uint32_t> new_point_ids(FRAC_UINT_INIT(vert_count * 2));
 
         // check the side of each vertex to the cutting plane
-        for (int i = 0; i < vert_count; i++) {
+        for (uint32_t i = 0; i < vert_count; i++) {
             vertices[i] = old_mesh.get_vertex(face.vert_ids[i]);
             side[i] = is_point_on_plane(p, n, vertices[i].pos) ? 0 : -1;
             side[i] = is_point_upside_plane(p, n, vertices[i].pos) ? 1 : -1;
